DeployedModel/requirements.txt
Flask==1.0
numpy==1.12.0
pandas==0.19.2
scikit-learn==0.18.1
scipy==0.18.1
Werkzeug==0.15.3
gunicorn==19.7.1
vaderSentiment==2.5
nltk==3.4.5
textstat==0.4.1
requests==2.20.0
flask_cors==3.0.9
cassandra-driver==3.24.0
uuid==1.30
